restructure vulkan helper class and function folders
