+++ include/vulkan/pipeline.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class Pipeline
+{
+public:
+
+private:
+};
