+++ src/material.cpp
@@ -5,7 +5,7 @@
 #define STB_IMAGE_IMPLEMENTATION
 #include <stb_image.h>
 
-#include "buffer.h"
+#include "vulkan/buffer.h"
 
 // --------------------------------------
 // TEXTURE POOL
