+++ src/vulkan/pipeline.cpp
@@ -1,0 +1,2 @@
+#include "vulkan/pipeline.h"
+
