+++ src/vulkan/vulkan_helpers.cpp
@@ -1,4 +1,4 @@
-#include "vulkan_helpers.h"
+#include "vulkan/vulkan_helpers.h"
 
 #include <algorithm>
 #include <fstream>
